Fix collision when hitbox crosses left/right side of the screen
This commit fixes a bug where an entity with a hitbox crossing the left
or right side of the screen can be erroneously pushed up or down by an
out-of-bounds solid tile.

There is no need to check for an overflow in
`_move_entity_and_collision_test_x_axis` as the tiles in the overflow
region will always have no collision.