Fix player hardlock with a watchdog timer
There is a bug in the code that causes the player's MetaSprite animation
to reset to 0 (walk_south) on a room transition.

This bug causes a hardlock if the player is knocked into another room by
an enemy.  The hurt state waits until the current animation ends, but it
never does, because walk_south is a looping animation.

This commit does not fix the room transition bug.  Instead it fixes the
hardlock with a watchdog timer that will reset the player's state when
the timer underflows.

Adding a watchdog to the player should reduce the risk of a future bug
causing another hardlock.