Remove automatic metasprite pattern matching
`find_best_pattern()` is too simple and doesn't find the best pattern.
I had to manually layout the dungeon MetaSprites to balance my tile
budget.

With the improved metasprite-editor GUI, manual layout is a lot easier.
`find_best_pattern()` is now a maintenance burden and so it has been
removed.

This commit also improves error messages when a frame is missing a
layout or references a missing pattern name.