Rewrite room loading code
This rewrite fixes a bug where the entities and room-events subsystems
were initialised and processed before the MetaSprite and MetaTile data
was loaded.

Room loading is now split in two.  First the room is copied to RAM.
Then, if required, MetaSprites/MetaTiles are loaded into memory.
Finally, the entities and room-events are initialised and processed.