Fix stale MetaTile graphics if room_event init() changes the map
This commit fixes a bug where the MetaTiles are drawn before the
room_event init() call in a fadein room transition.

This can cause an unlocked locked door to appear locked in a fadein room
transition.  The player can walk through the doorway as expected but
the on-screen graphics show locked door tiles.