Fix disappearing bomb when restoring a gamestate backup
If the player is holding a bomb, is hurt and the collision knock-back
knocks the player into a different room; the gamestate backup will state
the player is in the HURT state when the player should be in the
HOLDING_BOMB state.  If this erroneous gamestate backup is restored, the
player will not be holding a bomb.