Adjust the player's sword hitbox
Increased width of hitbox in attack_north and attack_south, so the
player can hit enemies slightly to the left (attack_north) or right
(attack_south) of the player.

Increased the size of attack_east0 and attack_west0 hitbox as I thought
it looked a little small.

Increased the height of attack_east1 and attack_west1 hitbox so the
hitbox's bottom matches the other attach_east and attack_west hitboxes.