Do not call reset or main
The S-SMP cannot be reset by S-CPU code.  As I do not know the state of
the audio-driver when `main()` is called, I am unable to reliably
reset and reinitialize the audio-driver on `audio.setup()`.

This commit removes the call to `reset()` in the break interrupt.  From
now on, if the code crashes the player must manually reset the console.

A double-underscore has been added to the reset/main function names to
prevent future-me from accidentally calling reset or main.