Improve metasprite tile deduplication
By adding all of the large tiles to the Tileset before processing any of
the small tiles.  This will deduplicate any small tile that also exists
in a large tile.

This change saves 7 small tiles in the common spritesheet and 1 small
tile in the dungeon spritesheet.

This change will also help simplify the upcoming dynamic sprite tile
subsystem.