Minor instrument cleanups
Removed unused instrument definitions (doesn't actually consume memory, but it's
redundant) and assigned ADSR to instruments instead of using the ADSR MML
command. This allows the parser to use a different command and save a byte.

My default setup is to use GF127 on the music instruments, since I generally
default to not using ADSR at all. Currently I have no desire to set default
ADSRs for the instruments instead of GAIN, but if the soundtrack is far enough
along, then the most common ADSR value can be embedded into the instrument,
saving two bytes per call... though currently that's not a common situation for
the soundtrack, because there's not a lot of instrument swaps that use ADSR.