Split second-layer into image & part_of_room namespaces and temporally remove part_of_room code
This commit moves the second-layer variables and code that is used when
the PART_OF_ROOM flag is clear to the second_layer.image namespace.

Some functions have been inlined and optimised as they are no-longer
required by PART_OF_ROOM.

This is in preparation for moving the second-layer PART_OF_ROOM map to
the room data.