Rewrite part_of_room second-layer to store the sl map data in room
I recently encountered a major downside my part_of_room second-layer -
I cannot select a room if the second-layer image is covering the room.
Clicking on the room will select the room with the second-layer, not the
room under the mouse cursor.  Hiding the second-layer image does not
select the room I want.

I've decided to store the part_of_room second-layer map in the room data
and continue to use the second-layer to draw the second-layer tilemap.
This will allow me to:
 * Reference 512 MetaTiles in a room with 2 layers
   (256 in the mt_tileset and 256 in the second-layer).
 * Allow second-layer reuse in dungeons with different mt_tilesets.

This changes does complicate the room Tiled UX a bit.  If a dungeon has
a part_of_room second-layer then:
 * the tmx file must contain 2 <layer> and 2 <tileset> tags
 * the mt_tileset <layer> must only use tiles in the mt_tileset <tileset>
 * the second-layer <layer> must only use tiles in the second-layer <tileset>

The <tileset> tag source attribute is used determine tmx <tileset>
belongs to which subsystem and the `SecondLayerInput.above_metatiles`
does the same for the <layer> tags.

To help simplify things, Tiled gid 0 will be mapped to tile 0 of both
tilesets.