Optimize Boss Theme
Some of the looping was not saving memory, instead increasing the filesize.
There is also at least one case where deeper nesting support was taken advantage
of, as there was a set of calls duplicated three times that previously wasn't
looped due to nesting limitations. Plus, in the cases of loops that did save
memory, it resulted in one of the subroutines being discarded because it no
longer saved memory to turn it into a subroutine.