Optimize Game Over Theme
Looping a single subroutine twice costs more memory than calling it twice, since
it's two bytes per call. Plus, in addition to some smaller loop cases, a
subroutine was discarded in favor of just using a loop with a breakpoint because
it was only called twice in a section that was close enough for the replacement.