Refactor to move gamestate vars to arrays with names in mappings.json
The gamestate variables are now stored in two arrays:
 1. A u8 array `gamestate.data.u8_vars : [ u8 ; ]`
 2. A u16 array ``

The elements in the array are named in the mappings.json file can now
be referenced by the compiler and wiz code (via gen/enums.wiz).

Gamestate variable access is assisted by functions (inline and jsr) in
`engine/gamestate_api.wiz`.

I decided to not add u32 variables to gamestate.  The only thing I could
see useful for u32 (in unnamed-snes-game) is time, and that could be
easily achieved with two u16 variables.