Further optimize Title Screen Theme
Some subroutines were in a loop that was only called twice then had a loop break
maker after the subroutine call. It is cheaper in memory to discard the loop
break marker and instead call the subroutine twice manually. An
instrument-related subroutine call was also put into another subroutine in order
to facilitate one of these kind of optimizations because it was the only other
call made beforehand. The entire song is also shortened in ticks because the
loop break optimization was causing an extra subroutine call to be added to the
very end: this is no longer necessary.