Use play pitch commands instead of samples in Rain Ambience + Thunder
This was done because these pitches currently aren't planned to be used
elsewhere. Therefore, by doing so, this will save common audio data bytes by
removing the sample entry (not to be confused with the instrument entry that
uses the same BRR file), and also save space in the common audio data pitch
table. It does consume a few extra bytes in the song file as a tradeoff, though.