Optimize Game Over Theme
The tempo slowdown effect was moved to channel H in order to merge some
notes and consolidate some extra notation into the subroutines. It does
result in the same effect in channel H with regards to adding slur markers
on the one note that is there as well as splitting up some rests, but the
overall result is saved bytes.