Update Factory Theme with an alternative pitch table for HardwareOverflowGlitch
A new instrument was created to facilitate the adaptation via a different pitch
table. Additional optimizations were done afterwards, including using
transposition in one place.

WARNING: This requires the lowest octave to be lowered by one in order to work
properly, as the beta release does not yet have this implemented.