Retire sine_P1O and update Final Dungeon Theme accordingly
This instrument was created to take full advantage of the octave range, but
the lowest octave is currently not in use. Subsequently, it has been
retired, as its only other use was in the Final Dungeon Theme, which now
uses transposition instead.

Note that this requires commit 4da96ccf204258608895452dfa18cb352be5b6db of
the Terrific Audio Driver at the earliest. This is because there is a note
that normally goes out of range in there due to the original instrument only
going up to octave 6, but ends up back in range because transposition takes
it down an octave.