Optimize Jungle Theme's long rest chains
Tie together some rests triggered multiple times in a row. This is because the
MML compiler splits the first two rests and merges all of the other ones, which
is fine when a key off is requested at the end of a rest like that off of a
slurred note... but these are already triggered once in a subroutine that ends
with a rest, meaning we only need one rest command generated.