Use a note range for HardwareOverflowGlitch_AltPitchTable
Nearly the entire third octave is not used, and this is the only instrument to
have a different set of pitch table entries. Given that it's only really used in
the Factory Theme and it's used to handle some really low notes, most of the
other notes will be excluded for now.

WARNING: This adaptation post-dates the beta2 release of v0.2.0.